requirements.txt
Django==3.0.5
gunicorn==20.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
model-mommy==2.0.0